% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fun.R
\name{clean_name}
\alias{clean_name}
\title{Clean name data}
\usage{
clean_name(to_remove, data_clean)
}
\arguments{
\item{to_remove}{Vector with characters to search and remove in data_clean}

\item{data_clean}{Vector suceptible to cleaning}
}
\value{
Second vector without characters from first vector and without ws
}
\description{
This function takes a character vector to remove and a vector to be cleaned.
Outputs the second vector without the characters in first vector and applies
trimws function
}
