% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_funs.R
\name{clean_roster}
\alias{clean_roster}
\title{Clean Rosters}
\usage{
clean_roster(seasons = 2021)
}
\arguments{
\item{seasons}{Range of years to pull from nflreadr rosters. Defaults to 2020}
}
\value{
Clean/prepared roster data
}
\description{
This function is a wrapper for nflreadr::load_rosters() and outputs all
headshot urls with https: prefix and integrates FB, HB and RB into RB.
Includes draft data
}
