% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.R
\name{fix_num}
\alias{fix_num}
\title{Fix number format in vector}
\usage{
fix_num(nums)
}
\arguments{
\item{nums}{Range of years to pull from nflreadr rosters. Defaults to 2020}
}
\value{
converted data to correct numeric form (eg. "20\%" to 0.2 and "2.3" to 2.3)
}
\description{
This function takes a vector and converts numbers or percentages into
numeric form while ignoring character columns for easy full data frame procesing.
(With dplyr::across for example)
}
