% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_funs.R
\name{foutsiders_data}
\alias{foutsiders_data}
\title{FOutsiders Login}
\usage{
foutsiders_data(page = "team-offense", season = 2020, user = "", pass = "")
}
\arguments{
\item{page}{page to scrape}

\item{season}{season to scrapte}

\item{user}{Football Outsiders authorized username}

\item{pass}{Football Outsiders authorized password}
}
\value{
table with requested data
}
\description{
This function takes a football ousiders page name, season as well as
credentials and outputs said table. Page name should be in the format
"team-offense", "basic-offensive-line", etc.
}
