% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_funs.R
\name{name_key}
\alias{name_key}
\title{Merge Name}
\usage{
name_key(name, draft_year, draft_round, position)
}
\arguments{
\item{name}{vector of NFL names to be converted to merge-ready format}

\item{draft_year}{vector of draft years in YYYY format}

\item{draft_round}{vector of single digit or character (1-7 or UDFA) indicating draft round}

\item{position}{position in uppercase format (eg. RB, WR)}
}
\value{
simplified name (using on ffscrapr mismatches, draft year, round and position)
}
\description{
This function takes a vector of NFL names and converts them to merge-ready format
}
