% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fun.R
\name{pull_AdvPass}
\alias{pull_AdvPass}
\title{Pull NFL Advanced Passing Stats}
\usage{
pull_AdvPass(seasons = 2020)
}
\arguments{
\item{seasons}{Range of years to pull from pfref. Defaults to 2020}
}
\value{
A data table including individual advanced passsing stats from
selected years
}
\description{
This function takes an argument for desired years and outputs a data.table
containing all Pro Football Reference individual advanced passing stats from selected years.
}
