% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fun.R
\name{pull_pass}
\alias{pull_pass}
\title{Pull NFL Passing Stats}
\usage{
pull_pass(seasons = 2020)
}
\arguments{
\item{seasons}{Range of years to pull from pfref. Defaults to 2020}
}
\value{
A data table including passsing stats from selected years
}
\description{
This function takes an argument for desired years and outputs a data.table
containing all Pro Football Reference passing stats from selected years.
}
