% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fun.R
\name{pull_scm}
\alias{pull_scm}
\title{Pull NFL Scrimmage Stats}
\usage{
pull_scm(seasons = 2020)
}
\arguments{
\item{seasons}{Range of years to pull from pfref. Defaults to 2020}
}
\value{
A data table including scrimmage stats from selected years
}
\description{
This function takes an argument for desired years and outputs a data.table
containing all Pro Football Reference scrimmage stats from selected years.
}
