% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fun.R
\name{pull_teamAdvPass}
\alias{pull_teamAdvPass}
\title{Pull NFL Advanced Team Passing Stats}
\usage{
pull_teamAdvPass(seasons = 2020)
}
\arguments{
\item{seasons}{Range of years to pull from pfref. Defaults to 2020}
}
\value{
A data table including team advancedpasssing stats from selected
years
}
\description{
This function takes an argument for desired years and outputs a data.table
containing all Pro Football Reference team advanced passing stats from selected years.
}
