% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.R
\name{standardize}
\alias{standardize}
\title{Standardize variable}
\usage{
standardize(var)
}
\arguments{
\item{var}{Numeric Variable}
}
\value{
Standardized variable
}
\description{
This function takes a vector of a numeric variable and converts them to a
standardized version of said variable
}
