% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.R
\name{theme_cas}
\alias{theme_cas}
\title{Custom ggplot theme}
\usage{
theme_cas(size = 13)
}
\arguments{
\item{size}{font size value (default = 13)}
}
\value{
ggplot2t heme
}
\description{
Applies a custom theme to ggplot object
}
