% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.R
\name{time_check}
\alias{time_check}
\title{Time Check}
\usage{
time_check(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\value{
time to run
}
\description{
Wrapper to check the time it takes to run a code snippet
}
