% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_funs.R
\name{name_key}
\alias{name_key}
\title{Merge Name}
\usage{
name_key(name, arg_1 = "", arg_2 = "", draft_year = F)
}
\arguments{
\item{name}{vector of NFL names to be converted to merge-ready format}

\item{arg_1}{(optional) additional argument to add specificity. (Draft Round recommended fo consistency reasons)}

\item{arg_2}{(optional) additional argument to add specificity. (position recommended fo consistency reasons)
#' @param draft_year (optional) vector of years in YYYY format}
}
\value{
simplified name (using on ffscrapr mismatches, draft year, and additional arguments)
}
\description{
This function takes a vector of NFL names and converts them to merge-ready format.
All arguments but name are optional and draft_year is optional but data entered must be in year format.
}
